+++ core/src/main/kotlin/co/infinum/collar/Collar.kt
@@ -20,6 +20,16 @@ object Collar {
     }
 
     /**
+     * Set the current analytics collection status (enabled or disabled).
+     *
+     * @param enabled is the analytics collection currently enabled.
+     */
+    @JvmStatic
+    fun setAnalyticsCollectionStatus(enabled: Boolean) {
+        collector?.setAnalyticsCollectionEnabled(enabled = enabled)
+    }
+
+    /**
      * Track screen names using a direct value.
      *
      * @param screenName value.